restructuring operators and manipulators on native types to its own module
