+++ tools/cmake/summary.cmake
@@ -95,6 +95,7 @@ function(universal_print_configuration_summary)
     universal_status("  BUILD_CI_CHECK               :   ${BUILD_CI_CHECK}")
     universal_status("")
     universal_status("  BUILD_BITBLOCK               :   ${BUILD_BITBLOCK}")
+    universal_status("  BUILD_NATIVE_TYPES           :   ${BUILD_NATIVE_TYPES}")
     universal_status("  BUILD_INTEGER_NUMBERS        :   ${BUILD_INTEGER_NUMBERS}")
     universal_status("  BUILD_DECIMAL_NUMBERS        :   ${BUILD_DECIMAL_NUMBERS}")
     universal_status("  BUILD_FIXPNT_NUMBERS         :   ${BUILD_FIXPNT_NUMBERS}")
